Consolidate and isolate bundle RPC integration tests
